requirements.txt
alyn==0.1.1
backports.functools-lru-cache==1.4
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
hunspell==0.5.0
matplotlib==2.1.0
networkx==2.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
pkg-resources==0.0.0
pyparsing==2.2.0
pytesseract==0.1.7
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
subprocess32==3.2.7

django
azure-storage
